pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "swarmee-river"
dynamic = ["version"]
description = "Swarmee River: an enterprise analytics + coding assistant built on the Strands Agents SDK."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Swarmee River"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "strands-agents[ollama]>=1.0.0",
    "strands-agents-tools>=0.1.0,<1.0.0",
    "strands-agents-sops>=1.0.6,<2.0.0",
    "boto3>=1.34.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
    "rich>=14.0.0,<15.0.0",
    "prompt_toolkit>=3.0.51,<4.0.0",
    "halo>=0.0.31,<1.0.0",
    "mypy>=1.0.0,<2.0.0",
    "pre-commit>=3.2.0,<4.2.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.0.0,<1.4.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-xdist>=3.0.0,<4.0.0",
    "ruff>=0.4.4,<1.0.0",
]

[project.scripts]
swarmee = "swarmee_river.swarmee:main"

[tool.hatch.metadata]
allow-direct-references = true

[project.urls]
Homepage = "https://github.com/strands-agents/agent-builder"
"Bug Tracker" = "https://github.com/strands-agents/agent-builder/issues"
Documentation = "https://strandsagents.com"

[project.optional-dependencies]
dev = [
    "commitizen>=4.4.0,<5.0.0",
    "hatch>=1.0.0,<2.0.0",
]
jupyter = [
    "ipykernel>=6.0.0,<7.0.0",
    "ipython>=8.0.0,<9.0.0",
    "nbformat>=5.0.0,<6.0.0",
]
docs = [
    "sphinx>=5.0.0,<6.0.0",
    "sphinx-rtd-theme>=1.0.0,<2.0.0",
    "sphinx-autodoc-typehints>=1.12.0,<2.0.0",
]
test = [
    "moto>=5.1.0,<6.0.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.4.0,<5.0.0",
    "hatch>=1.0.0,<2.0.0",
]

[tool.hatch.build]
packages = ["src/swarmee_river", "tools"]

[tool.hatch.envs.hatch-static-analysis]
dependencies = [
  "strands-agents[ollama]",
  "strands-agents-tools",
  "mypy>=1.0.0,<2.0.0",
  "ruff>=0.4.4,<1.0.0",
]

[tool.hatch.envs.hatch-static-analysis.scripts]
format-check = [
    "ruff format --check"
]
format-fix = [
    "ruff format"
]
lint-check = [
    "ruff check",
]
lint-fix = [
    "ruff check --fix"
]
type-check = [
    "mypy -p swarmee_river -p tools"
]

[tool.hatch.envs.hatch-test]
extra-dependencies = [
    "moto>=5.1.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.0.0,<1.4.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-xdist>=3.0.0,<4.0.0",
]
extra-args = [
    "-n",
    "auto",
    '-vv',
    '--no-header'
]

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.13", "3.12", "3.11", "3.10"]


[tool.hatch.envs.hatch-test.scripts]
run = [
    "pytest{env:HATCH_TEST_ARGS:} {args}"
]
run-cov = [
    "pytest{env:HATCH_TEST_ARGS:} --cov --cov-config=pyproject.toml {args}"
]

cov-combine = []
cov-report = []



[tool.hatch.envs.default.scripts]
list = [
    "echo 'Scripts commands available for default env:'; hatch env show --json | jq --raw-output '.default.scripts | keys[]'"
]
format = [
  "hatch fmt --formatter",
]
test-format = [
    "hatch fmt --formatter --check",
]
lint = [
    "hatch fmt --linter"
]
test-lint = [
    "hatch fmt --linter --check"
]
test = [
    "hatch test --cover --cov-report term --cov-report html --cov-report xml {args}"
]

test-integ = [
    "hatch test tests_integ {args}"
]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.mypy]
python_version = "3.10"
packages = ["swarmee_river", "tools"]
mypy_path = "src"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
follow_untyped_imports = true
ignore_missing_imports = false

[tool.ruff]
line-length = 120
include = ["src/**/*.py", "tests/**/*.py", "tools/**/*.py","tests_integ/**/*.py"]

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
]

[tool.coverage.run]
branch = true
source = ["src", "tools"]

[tool.coverage.report]
show_missing = true

[tool.coverage.html]
directory = "build/coverage/html"

[tool.coverage.xml]
output = "build/coverage/coverage.xml"

[tool.hatch.version]
# Tells Hatch to use your version control system (git) to determine the version.
source = "vcs"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
bump_message = "chore(release): bump version $current_version -> $new_version"
version_files = [
    "pyproject.toml:version",
]
update_changelog_on_bump = true
